# docs/swagger.yaml
definitions:
  handler.addFriendRequest:
    properties:
      friend_id:
        type: string
    type: object
...
  handler.errorEnvelope:
    properties:
      error:
        example: invalid email or password
        type: string
    type: object
...
  handler.friendListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.friendRow'
        type: array
    type: object
  handler.friendRow:
    properties:
      elo:
        type: integer
      handle:
        type: string
      id:
        type: string
      msg:
        type: string
      name:
        type: string
      presence:
        type: string
    type: object
...
  handler.giftRequest:
    properties:
      gift_id:
        type: string
    type: object
...
  handler.topupRequest:
    properties:
      pack_id:
        type: string
    type: object
  handler.topupResultEnvelope:
    properties:
      data:
        $ref: '#/definitions/service.TopupResult'
    type: object
  handler.txnListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
    type: object
...
  handler.userMini:
    properties:
      elo:
        type: integer
      handle:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  model.Transaction:
    properties:
      coins:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      kind:
        description: topup | gift
        type: string
      method:
        type: string
      status:
        description: ok | failed
        type: string
      vnd:
        type: integer
    type: object
...
  service.TopupResult:
    properties:
      coins:
        type: integer
      transaction:
        $ref: '#/definitions/model.Transaction'
    type: object
info:
  contact: {}
paths:
...
  /api/v1/friends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.friendListEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: List the authenticated user's friends
      tags:
      - friends
    post:
      consumes:
      - application/json
      parameters:
      - description: Friend id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.addFriendRequest'
      produces:
      - application/json
      responses:
        "204":
          description: added
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a friend
      tags:
      - friends
  /api/v1/friends/{id}:
    delete:
      parameters:
      - description: Friend user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: removed
      security:
      - BearerAuth: []
      summary: Remove a friend
      tags:
      - friends
...
  /api/v1/users/search:
    get:
      parameters:
      - description: name/email/handle
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.userMini'
              type: array
            type: object
      security:
      - BearerAuth: []
      summary: Search users to add as friends (excludes self & existing friends)
      tags:
      - friends
  /api/v1/wallet/gift:
    post:
      consumes:
      - application/json
      parameters:
      - description: Gift id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.giftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.topupResultEnvelope'
        "400":
          description: insufficient coins
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Send a gift (spend coins)
      tags:
      - wallet
  /api/v1/wallet/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Pack id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.topupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.topupResultEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Buy a coin pack (mock PayOS)
      tags:
      - wallet
  /api/v1/wallet/transactions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.txnListEnvelope'
      security:
      - BearerAuth: []
      summary: My wallet transactions
      tags:
      - wallet
swagger: "2.0"